+++ src/main/java/kitchenpos/application/TableService.java
@@ -4,9 +4,9 @@ import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.table.OrderTable;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
-import kitchenpos.ui.dto.ChangeNumberOfGuestsRequest;
-import kitchenpos.ui.dto.ChangeOrderTableEmptyRequest;
-import kitchenpos.ui.dto.CreateOrderTableRequest;
+import kitchenpos.dto.ChangeNumberOfGuestsRequest;
+import kitchenpos.dto.ChangeOrderTableEmptyRequest;
+import kitchenpos.dto.CreateOrderTableRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
