+++ src/main/java/kitchenpos/dto/ChangeOrderStatusRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.dto;
 
 public class ChangeOrderStatusRequest {
