+++ src/main/java/kitchenpos/dto/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.dto;
 
 import java.util.List;
