+++ src/main/java/kitchenpos/dto/CreateTableGroupOrderTableRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.dto;
 
 public class CreateTableGroupOrderTableRequest {
