+++ src/main/java/kitchenpos/dto/CreateTableGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.dto;
 
 import java.util.List;
 import java.util.stream.Collectors;
