+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.ui;
 
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.domain.menu.MenuGroup;
-import kitchenpos.ui.dto.CreateMenuGroupRequest;
+import kitchenpos.dto.CreateMenuGroupRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
