+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.ui;
 
 import kitchenpos.application.ProductService;
 import kitchenpos.domain.menu.Product;
-import kitchenpos.ui.dto.CreateProductRequest;
+import kitchenpos.dto.CreateProductRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
