+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -8,8 +8,8 @@ import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.repository.MenuRepository;
 import kitchenpos.repository.ProductRepository;
 import kitchenpos.domain.menu.MenuProducts;
-import kitchenpos.ui.dto.CreateMenuProductRequest;
-import kitchenpos.ui.dto.CreateMenuRequest;
+import kitchenpos.dto.CreateMenuProductRequest;
+import kitchenpos.dto.CreateMenuRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
