+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -15,9 +15,9 @@ import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 import kitchenpos.repository.ProductRepository;
 import kitchenpos.domain.menu.MenuProducts;
-import kitchenpos.ui.dto.ChangeOrderStatusRequest;
-import kitchenpos.ui.dto.CreateOrderLineItemRequest;
-import kitchenpos.ui.dto.CreateOrderRequest;
+import kitchenpos.dto.ChangeOrderStatusRequest;
+import kitchenpos.dto.CreateOrderLineItemRequest;
+import kitchenpos.dto.CreateOrderRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
