+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -8,9 +8,9 @@ import kitchenpos.domain.table.TableGroup;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 import kitchenpos.repository.TableGroupRepository;
-import kitchenpos.ui.dto.ChangeNumberOfGuestsRequest;
-import kitchenpos.ui.dto.ChangeOrderTableEmptyRequest;
-import kitchenpos.ui.dto.CreateOrderTableRequest;
+import kitchenpos.dto.ChangeNumberOfGuestsRequest;
+import kitchenpos.dto.ChangeOrderTableEmptyRequest;
+import kitchenpos.dto.CreateOrderTableRequest;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
